Fix jpg.cc for MinGW-w64
Fix the following error when building with MinGW-w64:

```raw
lib/extras/enc/jpg.cc:277:34: error: cannot convert 'long unsigned int*' to 'size_t*' {aka 'long long
unsigned int*'}
  277 |   jpeg_mem_dest(&cinfo, &buffer, &size);
      |                                  ^~~~~
      |                                  |
      |                                  long unsigned int*
```

+++ lib/extras/enc/jpg.cc
@@ -273,7 +273,7 @@ Status EncodeWithLibJpeg(const PackedImage& image, const JxlBasicInfo& info,
   cinfo.err = jpeg_std_error(&jerr);
   jpeg_create_compress(&cinfo);
   unsigned char* buffer = nullptr;
-  unsigned long size = 0;  // NOLINT
+  size_t size = 0;  // NOLINT
   jpeg_mem_dest(&cinfo, &buffer, &size);
   cinfo.image_width = image.xsize;
   cinfo.image_height = image.ysize;
